Renamed component from ProfileWallet to WalletOverview
